requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
pydantic==2.7.0
pydantic[email]==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
faker==18.6.0